pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["WEB"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^3.2"
djangorestframework = "^3.14.0"
coreapi = "^2.3.3"
requests = "^3.3.2"
apscheduler = "^3.10.1"
